requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Cors==4.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
qrcode[pil]==7.4.2
eventlet==0.36.1
Pillow==10.3.0
cryptography==42.0.8